feat:mudancas do tp-2

+++ src/main/java/com/projeto/bloco/gestao_fornecedores/repositories/FornecedorRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FornecedorRepository extends CrudRepository<Fornecedor, Integer> {
+public interface FornecedorRepository extends CrudRepository<Fornecedor, Long> {
 }
